Rewrite the DetectorMessenger
1. Remove declaration of G4UICommands. There were a lot of memeter.
So I desided to put them into map with holding a pointer to the data
associented with this command. The code succesfully shrinked.

+++ Config.h
@@ -38,7 +38,7 @@ const G4double mkrad = 1e-6*radian;
 
 struct Config_t
 {
-	unsigned photon_number=0; //number of photons in one pulse
+	double photon_number=0; //number of photons in one pulse
 
 	//geometry
     double world_size_x = 0.1*m;
